+++ SentryTestUtils/TestDispatchFactory.swift
@@ -5,6 +5,7 @@ import Foundation
 @_spi(Private) public class TestDispatchFactory: SentryDispatchFactory {
     public var vendedSourceHandler: ((TestDispatchSourceWrapper) -> Void)?
     public var vendedQueueHandler: ((TestSentryDispatchQueueWrapper) -> Void)?
+    public var vendedUtilityQueueHandler: ((TestSentryDispatchQueueWrapper) -> Void)?
 
     public var createUtilityQueueInvocations = Invocations<(name: String, relativePriority: Int32)>()
 
@@ -18,7 +19,9 @@ import Foundation
         createUtilityQueueInvocations.record((String(cString: name), relativePriority))
         // Due to the absense of `dispatch_queue_attr_make_with_qos_class` in Swift, we do not pass any attributes.
         // This will not affect the tests as they do not need an actual low priority queue.
-        return TestSentryDispatchQueueWrapper(name: name, attributes: nil)
+        let queue = TestSentryDispatchQueueWrapper(name: name, attributes: nil)
+        vendedUtilityQueueHandler?(queue)
+        return queue
     }
 
     public override func source(withInterval interval: Int, leeway: Int, queueName: UnsafePointer<CChar>, attributes: __OS_dispatch_queue_attr, eventHandler: @escaping () -> Void) -> SentryDispatchSourceWrapper {