+++ Sources/Swift/Tools/SentryLogger.swift
@@ -216,10 +216,6 @@ public final class SentryLogger: NSObject {
         }
         
         if let processedLog {
-            SentrySDKLog.log(
-                message: "[SentryLogger] \(processedLog.body)",
-                andLevel: processedLog.level.toSentryLevel()
-            )
             batcher.add(processedLog)
         }
     }